Fixed the web routes. Added some logs to make it easier to debug when doing stress testing.

+++ order-service/source/database/migrations/2025_02_14_180459_create_orders_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
 
-            $table->unsignedTinyInteger("user_id");
+            $table->unsignedBigInteger("user_id");
             $table->enum("status", ["open","pending","completed","cancelled"]);
 
             $table->timestamps();
